+++ packages/react-reshaped/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @dfosco/storyboard-react-reshaped
 
+## 3.10.0-beta.1
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @dfosco/storyboard-react@3.10.0-beta.1
+
+## 3.10.0-beta.0
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @dfosco/storyboard-react@3.10.0-beta.0
+
 ## 3.9.1
 
 ### Patch Changes